Added ComponentArray which requires each component to have the same constructor interface. Hence, all components now have as a constructor input parameters only 'name'. Input has now an extra constructor instead of replacing the old ones requiring input value(s). The old constructors can be used when Input is not in ComponentArray. Hence, Input::makeSignal() has been added to use Input when created using the new constructor. Parameters referring to pin numbers are now of type unsigned short int (ushort). Component numOfInPins, numOfOutPins, and allInReached are now replaced by getters for encapsulation. Parent out update has been removed. It seems unnecessary because the out variable cannot be accessed. Out vales are returned by the getOutVal() or getOutVec() functions which do not need to have updated value of 'out' of the parents - only children use this value. In verbose mode, wires prompt also the value they are propagating, not only components and pins. Added DCSHeader.h file (not sure how to use it) Added parent's name to Node naming inside DFlipFlop.

+++ Digital circuit simulator/Connections/DCSWire.hpp
@@ -19,22 +19,22 @@ private:
 	// Driving component
 	DCSComponent* from;
 	// output pin number of the output component
-	int outPinNum;
+	ushort outPinNum;
 	// Drived component
 	DCSComponent* to;
 	// input pin number of the driven component
-	int inPinNum;
+	ushort inPinNum;
 	std::string probeName;
 	
 public:
 	DCSWire() = delete;
 	DCSWire(DCSComponent* from,
-			int outPinNum,
+			ushort outPinNum,
 			DCSComponent* to,
-			int inPinNum,
+			ushort inPinNum,
 			std::string probeName = "");
 	
-	int getOutPinNum();
+	ushort getOutPinNum();
 	std::string getProbeName();
 	void propagateValue();
 	bool getOutVal();
